+++ tw2/core/validation.py
@@ -441,7 +441,7 @@ class DateTimeValidator(RangeValidator):
         'toosmall': ('date_toosmall', _('Cannot be earlier than $min_str')),
         'toobig': ('date_toobig', _('Cannot be later than $max_str')),
     }
-    format = '%d/%m/%Y %H:%M'
+    format = '%Y-%m-%d %H:%M'
 
     format_tbl = {
         'd': 'day',
@@ -494,7 +494,7 @@ class DateValidator(DateTimeValidator):
         'baddatetime': (
             'baddate', _('Must follow date format $format_str')),
     }
-    format = '%d/%m/%Y'
+    format = '%Y-%m-%d'
 
     def _convert_to_python(self, value, state=None):
         value = super(DateValidator, self)._convert_to_python(value)
@@ -602,25 +602,44 @@ class MatchValidator(Validator):
 
     `other_field`
         Name of the sibling field this must match
+    `pass_on_invalid`
+        Pass validation if sibling field is Invalid
     """
     msgs = {
         'mismatch': _("Must match $other_field_str"),
+        'notfound': _("$other_field_str field is not found"),
+        'invalid': _("$other_field_str field is invalid"),
     }
 
-    def __init__(self, other_field, **kw):
+    def __init__(self, other_field, pass_on_invalid=False, **kw):
         super(MatchValidator, self).__init__(**kw)
         self.other_field = other_field
+        self.pass_on_invalid = pass_on_invalid
 
     @property
     def other_field_str(self):
         return capitalize(util.name2label(self.other_field).lower())
 
     def _validate_python(self, value, state):
-        if self.other_field not in state or value != state[self.other_field]:
+        if isinstance(state, dict):
+            # Backward compatibility
+            values = state
+        else:
+            values = state.full_dict
+
+        if self.other_field not in values:
+            raise ValidationError('notfound', self)
+
+        other_value = values[self.other_field]
+
+        if other_value is Invalid:
+            if not self.pass_on_invalid:
+                raise ValidationError('invalid', self)
+        elif value != other_value:
             raise ValidationError('mismatch', self)
 
     def _is_empty(self, value):
-        return self.required and super(MatchValidator, self)._is_empty
+        return self.required and super(MatchValidator, self)._is_empty(value)
 
 
 class CompoundValidator(Validator):